backend/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.0
python-multipart>=0.0.6
google-generativeai>=0.1.0
pinecone>=2.2.1  # Changed from pinecone-client to pinecone
python-dotenv>=1.0.0
aiohttp>=3.8.4
pydantic>=1.10.7
PyPDF2==3.0.1
python-docx==0.8.11
openai
